+++ pykeg/web/kegweb/views.py
@@ -116,7 +116,11 @@ class KegListView(ListView):
 
 def fullscreen(request):
     context = RequestContext(request)
-    context['taps'] = models.KegTap.objects.all()
+    taps = models.KegTap.objects.all()
+    active_taps = [t for t in taps if t.current_keg]
+    pages = [active_taps[i:i+4] for i in range(0, len(active_taps), 4)]
+    context['pages'] = pages
+
     return render_to_response('kegweb/fullscreen.html', context_instance=context)