+++ docs/source/changelog.rst
@@ -5,6 +5,13 @@ Changelog
 
 **Upgrade Procedure:** Please follow :ref:`upgrading-kegbot` for general upgrade steps.
 
+Current Version (in development)
+--------------------------------
+* API: New endpoint: `drinks/last`.
+* Newly-created meters now default to FT330-RJ calibration values.
+* Kegadmin: Kegs can be deleted from the "Edit Keg" screen.
+
+
 Version 1.1.0 (2014-09-19)
 --------------------------
 * Fullscreen mode.